===BEGIN add-rollup=== Rename babel-loader to babel
